refactor: api.ts 이름 및 디렉토리 구조 변경
